Enum classes are broken #3

+++ include/kaguya/type.hpp
@@ -301,11 +301,11 @@ namespace kaguya
 		}
 		static get_type get(lua_State* l, int index)
 		{
-			return static_cast<get_type>(lua_type_traits<int>::get(l, index));
+			return static_cast<get_type>(lua_type_traits<int64_t>::get(l, index));
 		}
 		static int push(lua_State* l, push_type s)
 		{
-			return lua_type_traits<int>::push(l, s);
+			return lua_type_traits<int64_t>::push(l, static_cast<int64_t>(s));
 		}
 	};
